Added random message from actions.json file

+++ src/sql/character.sql
@@ -1,5 +1,5 @@
 CREATE TABLE characters (
-    user_id INT UNIQUE PRIMARY KEY,
+    user_id BIGINT UNIQUE PRIMARY KEY,
     registered DATETIME DEFAULT CURRENT_TIMESTAMP,
     name VARCHAR(64) NOT NULL,
     element INTEGER,
